Demo Polls App
All the 32 files of the polls app are included in the repository. We can use this as a basic skeleton to start with our project. The GUI sucks very well so if anyone wants to modify the look..feel free to branch. Also download this and run this webapp on your machines.
The credentials for admin site are as follows : 
Username : shravan
password : iamgroot